Halfway implemented timers
They way I have done it isn't very good. It would be much better to put the
timers in as "fake" incursions that show when they despawned and when they
should respawn. I should also have some timers that show when the incursion is
going to go into withdrawing from mobilizing.